+++ web/src/settings_data.ts
@@ -192,19 +192,9 @@ export function can_manage_user_group(group_id: number): boolean {
         return false;
     }
 
-    let can_manage_all_groups = user_can_manage_all_groups();
-
     const group = user_groups.get_user_group_from_id(group_id);
 
-    if (
-        !current_user.is_admin &&
-        !current_user.is_moderator &&
-        !user_groups.is_direct_member_of(current_user.user_id, group_id)
-    ) {
-        can_manage_all_groups = false;
-    }
-
-    if (can_manage_all_groups) {
+    if (user_can_manage_all_groups()) {
         return true;
     }
 
@@ -213,6 +203,22 @@ export function can_manage_user_group(group_id: number): boolean {
         "can_manage_group",
         "group",
     );
+}
+
+export function can_add_members_to_user_group(group_id: number): boolean {
+    const group = user_groups.get_user_group_from_id(group_id);
+
+    if (
+        user_has_permission_for_group_setting(
+            group.can_add_members_group,
+            "can_add_members_group",
+            "group",
+        )
+    ) {
+        return true;
+    }
+
+    return can_manage_user_group(group_id);
 }
 
 export function can_join_user_group(group_id: number): boolean {