+++ web/src/user_group_create.ts
@@ -28,6 +28,7 @@ export function get_name(): string | undefined {
     return created_group_name;
 }
 
+let can_add_members_group_widget: GroupSettingPillContainer | undefined;
 let can_manage_group_widget: GroupSettingPillContainer | undefined;
 let can_join_group_widget: GroupSettingPillContainer | undefined;
 
@@ -154,6 +155,11 @@ function create_user_group(): void {
     }
     const user_ids = user_group_create_members.get_principals();
 
+    assert(can_add_members_group_widget !== undefined);
+    const can_add_members_group = settings_components.get_group_setting_widget_value(
+        can_add_members_group_widget,
+    );
+
     assert(can_manage_group_widget !== undefined);
     const can_manage_group =
         settings_components.get_group_setting_widget_value(can_manage_group_widget);
@@ -174,6 +180,7 @@ function create_user_group(): void {
         name: group_name,
         description,
         members: JSON.stringify(user_ids),
+        can_add_members_group: JSON.stringify(can_add_members_group),
         can_join_group: JSON.stringify(can_join_group),
         can_mention_group,
         can_manage_group: JSON.stringify(can_manage_group),
@@ -246,6 +253,12 @@ export function set_up_handlers(): void {
         if (keydown_util.is_enter_event(e)) {
             e.preventDefault();
         }
+    });
+
+    can_add_members_group_widget = settings_components.create_group_setting_widget({
+        $pill_container: $container.find(".can-add-members-group-container .pill-container"),
+        setting_name: "can_add_members_group",
+        setting_type: "group",
     });
 
     const $pill_container = $container.find(".can-manage-group-container .pill-container");