+++ zerver/lib/user_groups.py
@@ -55,6 +55,7 @@ class UserGroupDict(TypedDict):
     creator_id: int | None
     date_created: int | None
     is_system_group: bool
+    can_add_members_group: int | AnonymousSettingGroupDict
     can_join_group: int | AnonymousSettingGroupDict
     can_manage_group: int | AnonymousSettingGroupDict
     can_mention_group: int | AnonymousSettingGroupDict
@@ -131,24 +132,6 @@ def access_user_group_to_read_membership(user_group_id: int, realm: Realm) -> Na
     return get_user_group_by_id_in_realm(user_group_id, realm, for_read=True)
 
 
-def check_permission_for_managing_all_groups(
-    user_group: UserGroup, user_profile: UserProfile
-) -> bool:
-    """
-    Given a user and a group in the same realm, checks if the user
-    can manage the group through the legacy can_manage_all_groups
-    permission, which is a permission that requires either certain roles
-    or membership in the group itself to be used.
-    """
-    can_manage_all_groups = user_profile.can_manage_all_groups()
-    if can_manage_all_groups:
-        if user_profile.is_realm_admin or user_profile.is_moderator:
-            return True
-
-        return is_user_in_group(user_group, user_profile)
-    return False
-
-
 def access_user_group_for_update(
     user_group_id: int,
     user_profile: UserProfile,
@@ -172,8 +155,13 @@ def access_user_group_for_update(
         raise JsonableError(_("Insufficient permission"))
 
     assert permission_setting in NamedUserGroup.GROUP_PERMISSION_SETTINGS
-    if permission_setting == "can_manage_group" and check_permission_for_managing_all_groups(
-        user_group, user_profile
+    if (
+        permission_setting
+        in [
+            "can_manage_group",
+            "can_add_members_group",
+        ]
+        and user_profile.can_manage_all_groups()
     ):
         return user_group
 
@@ -182,6 +170,17 @@ def access_user_group_for_update(
         user_profile,
         NamedUserGroup.GROUP_PERMISSION_SETTINGS[permission_setting],
     )
+
+    # Users with permission to manage the group should be able to add members
+    # to the group. This also takes care of the case where a user creating a group
+    # with themselves having the permission to manage it don't have to explicitly
+    # add themselves to can_add_members_group.
+    if not user_has_permission and permission_setting == "can_add_members_group":
+        user_has_permission = user_has_permission_for_group_setting(
+            user_group.can_manage_group,
+            user_profile,
+            NamedUserGroup.GROUP_PERMISSION_SETTINGS["can_manage_group"],
+        )
 
     if not user_has_permission:
         raise JsonableError(_("Insufficient permission"))
@@ -554,6 +553,8 @@ def user_groups_in_realm_serialized(
     UserGroup and UserGroupMembership that we need.
     """
     realm_groups = NamedUserGroup.objects.select_related(
+        "can_add_members_group",
+        "can_add_members_group__named_user_group",
         "can_join_group",
         "can_join_group__named_user_group",
         "can_manage_group",
@@ -610,6 +611,9 @@ def user_groups_in_realm_serialized(
             members=direct_member_ids,
             direct_subgroup_ids=direct_subgroup_ids,
             is_system_group=user_group.is_system_group,
+            can_add_members_group=get_setting_value_for_user_group_object(
+                user_group.can_add_members_group, group_members, group_subgroups
+            ),
             can_join_group=get_setting_value_for_user_group_object(
                 user_group.can_join_group, group_members, group_subgroups
             ),
@@ -834,12 +838,13 @@ def bulk_create_system_user_groups(groups: list[dict[str, str]], realm: Realm) -
         user_group_ids = [id for (id,) in cursor.fetchall()]
 
     rows = [
-        SQL("({},{},{},{},{},{},{},{},{})").format(
+        SQL("({},{},{},{},{},{},{},{},{},{})").format(
             Literal(user_group_ids[idx]),
             Literal(realm.id),
             Literal(group["name"]),
             Literal(group["description"]),
             Literal(True),
+            Literal(initial_group_setting_value),
             Literal(initial_group_setting_value),
             Literal(initial_group_setting_value),
             Literal(initial_group_setting_value),
@@ -849,7 +854,7 @@ def bulk_create_system_user_groups(groups: list[dict[str, str]], realm: Realm) -
     ]
     query = SQL(
         """
-        INSERT INTO zerver_namedusergroup (usergroup_ptr_id, realm_id, name, description, is_system_group, can_join_group_id, can_manage_group_id, can_mention_group_id, deactivated)
+        INSERT INTO zerver_namedusergroup (usergroup_ptr_id, realm_id, name, description, is_system_group, can_add_members_group_id, can_join_group_id, can_manage_group_id, can_mention_group_id, deactivated)
         VALUES {rows}
         """
     ).format(rows=SQL(", ").join(rows))
@@ -931,7 +936,8 @@ def create_system_user_groups_for_realm(realm: Realm) -> dict[int, NamedUserGrou
         user_group = set_defaults_for_group_settings(group, {}, system_groups_name_dict)
         groups_with_updated_settings.append(user_group)
     NamedUserGroup.objects.bulk_update(
-        groups_with_updated_settings, ["can_join_group", "can_manage_group", "can_mention_group"]
+        groups_with_updated_settings,
+        ["can_add_members_group", "can_join_group", "can_manage_group", "can_mention_group"],
     )
 
     subgroup_objects: list[GroupGroupMembership] = []