+++ api_docs/changelog.md
@@ -20,6 +20,18 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 10.0
 
+**Feature level 304**
+
+* [`POST /register`](/api/register-queue), [`GET /events`](/api/get-events),
+  [`GET /user_groups`](/api/get-user-groups): Add `can_add_members_group` to
+  user group objects.
+* [`POST /user_groups/create`](/api/create-user-group): Added `can_add_members_group`
+  parameter to support setting the user group which can add members to the user
+  group.
+* [`PATCH /user_groups/{user_group_id}`](/api/update-user-group): Added
+  `can_add_members_group` parameter to support changing the user group which
+  can add members to the specified user group.
+
 **Feature level 303**
 
 * [`POST /register`](/api/register-queue), [`GET /user_groups`](/api/get-user-groups),